pyproject.toml
[project]
name = "pdf2text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "azure-ai-formrecognizer>=3.3.3",
    "azure-core>=1.37.0",
    "azure-identity>=1.25.1",
    "lagom>=2.7.7",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tabulate>=0.9.0",
]

[dependency-groups]
dev = [
    "pip-audit>=2.10.0",
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.11",
]

[tool.ruff]
exclude = ["typings", ".venv"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", ".vscode", ".git"]

[tool.coverage.report]
fail_under = 90

[tool.coverage.run]
omit = [
    "*/__init__.py",
    "pdf2text/protocols/*.py",
    "pdf2text/hosting.py"
]

